+++ src/vector.c
@@ -70,7 +70,7 @@ int igraph_vector_floor(const igraph_vector_t *from, igraph_vector_long_t *to) {
   
   IGRAPH_CHECK(igraph_vector_long_resize(to, n));
   for (i=0; i<n; i++) {
-    VECTOR(*to)[i] = floor(VECTOR(*from)[i]);
+    VECTOR(*to)[i] = (long int) floor(VECTOR(*from)[i]);
   }
   return 0;
 }
@@ -80,7 +80,7 @@ int igraph_vector_round(const igraph_vector_t *from, igraph_vector_long_t *to) {
   
   IGRAPH_CHECK(igraph_vector_long_resize(to, n));
   for (i=0; i<n; i++) {
-    VECTOR(*to)[i] = round(VECTOR(*from)[i]);
+    VECTOR(*to)[i] = (long int) round(VECTOR(*from)[i]);
   }
   return 0;
 }
@@ -134,12 +134,12 @@ int igraph_vector_order(const igraph_vector_t* v,
   assert(v!=NULL);
   assert(v->stor_begin != NULL);
 
-  IGRAPH_VECTOR_INIT_FINALLY(&ptr, nodes+1);
+  IGRAPH_VECTOR_INIT_FINALLY(&ptr, (long int) nodes+1);
   IGRAPH_VECTOR_INIT_FINALLY(&rad, edges);
   IGRAPH_CHECK(igraph_vector_resize(res, edges));
   
   for (i=0; i<edges; i++) {
-    long int radix=v2->stor_begin[i];
+    long int radix=(long int) v2->stor_begin[i];
     if (VECTOR(ptr)[radix]!=0) {
       VECTOR(rad)[i]=VECTOR(ptr)[radix];
     }
@@ -149,10 +149,10 @@ int igraph_vector_order(const igraph_vector_t* v,
   j=0;
   for (i=0; i<nodes+1; i++) {
     if (VECTOR(ptr)[i] != 0) {
-      long int next=VECTOR(ptr)[i]-1;
+      long int next=(long int) VECTOR(ptr)[i]-1;
       res->stor_begin[j++]=next;
       while (VECTOR(rad)[next] != 0) {
-	next=VECTOR(rad)[next]-1;
+	next=(long int) VECTOR(rad)[next]-1;
 	res->stor_begin[j++]=next;
       }
     }
@@ -162,8 +162,8 @@ int igraph_vector_order(const igraph_vector_t* v,
   igraph_vector_null(&rad);
 
   for (i=0; i<edges; i++) {
-    long int edge=VECTOR(*res)[edges-i-1];
-    long int radix=VECTOR(*v)[edge];
+    long int edge=(long int) VECTOR(*res)[edges-i-1];
+    long int radix=(long int) VECTOR(*v)[edge];
     if (VECTOR(ptr)[radix]!= 0) {
       VECTOR(rad)[edge]=VECTOR(ptr)[radix];
     }
@@ -173,10 +173,10 @@ int igraph_vector_order(const igraph_vector_t* v,
   j=0;
   for (i=0; i<nodes+1; i++) {
     if (VECTOR(ptr)[i] != 0) {
-      long int next=VECTOR(ptr)[i]-1;
+      long int next=(long int) VECTOR(ptr)[i]-1;
       res->stor_begin[j++]=next;
       while (VECTOR(rad)[next] != 0) {
-	next=VECTOR(rad)[next]-1;
+	next=(long int) VECTOR(rad)[next]-1;
 	res->stor_begin[j++]=next;
       }
     }
@@ -199,12 +199,12 @@ int igraph_vector_order1(const igraph_vector_t* v,
   assert(v!=NULL);
   assert(v->stor_begin != NULL);
 
-  IGRAPH_VECTOR_INIT_FINALLY(&ptr, nodes+1);
+  IGRAPH_VECTOR_INIT_FINALLY(&ptr, (long int) nodes+1);
   IGRAPH_VECTOR_INIT_FINALLY(&rad, edges);
   IGRAPH_CHECK(igraph_vector_resize(res, edges));
   
   for (i=0; i<edges; i++) {
-    long int radix=v->stor_begin[i];
+    long int radix=(long int) v->stor_begin[i];
     if (VECTOR(ptr)[radix]!=0) {
       VECTOR(rad)[i]=VECTOR(ptr)[radix];
     }
@@ -214,10 +214,10 @@ int igraph_vector_order1(const igraph_vector_t* v,
   j=0;
   for (i=0; i<nodes+1; i++) {
     if (VECTOR(ptr)[i] != 0) {
-      long int next=VECTOR(ptr)[i]-1;
+      long int next=(long int) VECTOR(ptr)[i]-1;
       res->stor_begin[j++]=next;
       while (VECTOR(rad)[next] != 0) {
-	next=VECTOR(rad)[next]-1;
+	next=(long int) VECTOR(rad)[next]-1;
 	res->stor_begin[j++]=next;
       }
     }
@@ -243,16 +243,16 @@ int igraph_vector_rank(const igraph_vector_t *v, igraph_vector_t *res,
   IGRAPH_CHECK(igraph_vector_resize(res, edges));
 	       
   for (i=0; i<edges; i++) {
-    long int elem=VECTOR(*v)[i];
+    long int elem=(long int) VECTOR(*v)[i];
     VECTOR(ptr)[i] = VECTOR(rad)[elem];
     VECTOR(rad)[elem] = i+1;
   }
   
   for (i=0; i<nodes; i++) {
-    long int p=VECTOR(rad)[i];
+    long int p=(long int) VECTOR(rad)[i];
     while (p != 0) {      
       VECTOR(*res)[p-1]=c++;
-      p=VECTOR(ptr)[p-1];
+      p=(long int) VECTOR(ptr)[p-1];
     }
   }
 
@@ -295,7 +295,7 @@ int igraph_vector_complex_fprint(const igraph_vector_complex_t *v,
 
 int igraph_vector_complex_real(const igraph_vector_complex_t *v, 
 			       igraph_vector_t *real) {
-  int i, n=igraph_vector_complex_size(v);
+  int i, n=(int) igraph_vector_complex_size(v);
   IGRAPH_CHECK(igraph_vector_resize(real, n));
   for (i=0; i<n; i++) {
     VECTOR(*real)[i] = IGRAPH_REAL(VECTOR(*v)[i]);
@@ -306,7 +306,7 @@ int igraph_vector_complex_real(const igraph_vector_complex_t *v,
 
 int igraph_vector_complex_imag(const igraph_vector_complex_t *v, 
 			       igraph_vector_t *imag) {
-  int i, n=igraph_vector_complex_size(v);
+  int i, n=(int) igraph_vector_complex_size(v);
   IGRAPH_CHECK(igraph_vector_resize(imag, n));
   for (i=0; i<n; i++) {
     VECTOR(*imag)[i] = IGRAPH_IMAG(VECTOR(*v)[i]);
@@ -318,7 +318,7 @@ int igraph_vector_complex_imag(const igraph_vector_complex_t *v,
 int igraph_vector_complex_realimag(const igraph_vector_complex_t *v, 
 				   igraph_vector_t *real, 
 				   igraph_vector_t *imag) {
-  int i, n=igraph_vector_complex_size(v);
+  int i, n=(int) igraph_vector_complex_size(v);
   IGRAPH_CHECK(igraph_vector_resize(real, n));
   IGRAPH_CHECK(igraph_vector_resize(imag, n));
   for (i=0; i<n; i++) {
@@ -333,7 +333,7 @@ int igraph_vector_complex_realimag(const igraph_vector_complex_t *v,
 int igraph_vector_complex_create(igraph_vector_complex_t *v,
 				 const igraph_vector_t *real,
 				 const igraph_vector_t *imag) {
-  int i, n=igraph_vector_size(real);
+  int i, n=(int) igraph_vector_size(real);
   if (n != igraph_vector_size(imag)) {
     IGRAPH_ERROR("Real and imag vector sizes don't match", IGRAPH_EINVAL);
   }
@@ -351,7 +351,7 @@ int igraph_vector_complex_create(igraph_vector_complex_t *v,
 int igraph_vector_complex_create_polar(igraph_vector_complex_t *v,
 				       const igraph_vector_t *r,
 				       const igraph_vector_t *theta) {
-  int i, n=igraph_vector_size(r);
+  int i, n=(int) igraph_vector_size(r);
   if (n != igraph_vector_size(theta)) {
     IGRAPH_ERROR("'r' and 'theta' vector sizes don't match", IGRAPH_EINVAL);
   }
